разделил тесты по разным файлам с подходящими названиями соответствующие контексту тестов
